# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PrejuZero API",
    "version": "1",
    "description": "API do PrejuZero. Erros em application/problem+json (RFC 9457), paginação por cursor e Idempotency-Key em operações sensíveis (ADR-009)."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "paths": {
...
    "/v1/auth/2fa/ativar": {
      "post": {
        "operationId": "ativarSegundoFator",
        "summary": "Ativa o 2FA e devolve os códigos de recuperação.",
        "tags": [
          "auth"
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodigoSegundoFator"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SegundoFatorAtivado"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "409": {
            "description": "Conflito com o estado atual.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "500": {
            "description": "Erro inesperado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/2fa/configurar": {
      "post": {
        "operationId": "configurarSegundoFator",
        "summary": "Gera o segredo TOTP e a URI do QR code.",
        "tags": [
          "auth"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConfiguracaoSegundoFator"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "409": {
            "description": "Conflito com o estado atual.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "500": {
            "description": "Erro inesperado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/2fa/verificar": {
      "post": {
        "operationId": "verificarSegundoFator",
        "summary": "Verifica o código do aplicativo ou de recuperação.",
        "tags": [
          "auth"
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CodigoSegundoFator"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessaoAtual"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "500": {
            "description": "Erro inesperado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/acessos": {
      "get": {
        "operationId": "listarAcessos",
        "summary": "Últimos acessos da conta (logins, 2FA, saídas e bloqueios).",
        "tags": [
          "auth"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AcessosRecentes"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "500": {
            "description": "Erro inesperado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/dispositivos": {
      "get": {
        "operationId": "listarDispositivos",
        "summary": "Dispositivos com sessão na conta.",
        "tags": [
          "auth"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DispositivosDaConta"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "500": {
            "description": "Erro inesperado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/dispositivos/{id}": {
      "delete": {
        "operationId": "revogarDispositivo",
        "summary": "Encerra a sessão de um dispositivo na hora.",
        "tags": [
          "auth"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Sucesso, sem corpo."
          },
          "400": {
            "description": "Entrada inválida.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "404": {
            "description": "Não encontrado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "500": {
            "description": "Erro inesperado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/entrar": {
      "post": {
        "operationId": "entrar",
        "summary": "Entra com e-mail e senha; a sessão nasce aguardando o 2FA.",
        "tags": [
          "auth"
        ],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Credenciais"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessaoAtual"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "429": {
            "description": "Limite de requisições excedido.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "500": {
            "description": "Erro inesperado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/eu": {
      "get": {
        "operationId": "consultarSessao",
        "summary": "Usuário, tenant e nível da sessão atual.",
        "tags": [
          "auth"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessaoAtual"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "500": {
            "description": "Erro inesperado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/sair": {
      "post": {
        "operationId": "sair",
        "summary": "Encerra a sessão atual.",
        "tags": [
          "auth"
        ],
        "parameters": [],
        "responses": {
          "204": {
            "description": "Sucesso, sem corpo."
          },
          "401": {
            "description": "Não autenticado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "500": {
            "description": "Erro inesperado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/senha/esqueci": {
      "post": {
        "operationId": "solicitarRedefinicaoDeSenha",
        "summary": "Envia o link de redefinição de senha, se o e-mail tiver conta.",
        "tags": [
          "auth"
        ],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PedidoDeRedefinicaoDeSenha"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Sucesso, sem corpo."
          },
          "400": {
            "description": "Entrada inválida.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "429": {
            "description": "Limite de requisições excedido.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "500": {
            "description": "Erro inesperado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/senha/redefinir": {
      "post": {
        "operationId": "redefinirSenha",
        "summary": "Define a nova senha com o token recebido por e-mail.",
        "tags": [
          "auth"
        ],
        "security": [],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RedefinicaoDeSenha"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Sucesso, sem corpo."
          },
          "400": {
            "description": "Entrada inválida.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "429": {
            "description": "Limite de requisições excedido.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "500": {
            "description": "Erro inesperado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/tokens": {
      "post": {
        "operationId": "emitirTokensDeDispositivo",
        "summary": "Registra o dispositivo e emite token de acesso curto e de renovação.",
        "tags": [
          "auth"
        ],
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PedidoDeTokensDeDispositivo"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Sucesso.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokensDeDispositivo"
                }
              }
            }
          },
          "400": {
            "description": "Entrada inválida.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "401": {
            "description": "Não autenticado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "403": {
            "description": "Sem permissão.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          },
          "500": {
            "description": "Erro inesperado.",
            "content": {
              "application/problem+json": {
                "schema": {
                  "$ref": "#/components/schemas/Problema"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CodigoSegundoFator": {
        "type": "object",
        "properties": {
          "codigo": {
            "type": "string",
            "minLength": 6,
            "maxLength": 16,
            "description": "6 dígitos do aplicativo ou um código de recuperação (XXXXX-XXXXX)."
          }
        },
        "required": [
          "codigo"
        ],
        "additionalProperties": false
      },
...
      "PedidoDeRedefinicaoDeSenha": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "minLength": 1,
            "maxLength": 254
          }
        },
        "required": [
          "email"
        ],
        "additionalProperties": false
      },
...
      "Problema": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "description": "URI que identifica o tipo do problema."
          },
          "title": {
            "type": "string",
            "description": "Resumo legível, em pt-BR."
          },
          "status": {
            "type": "integer",
            "minimum": 400,
            "maximum": 599
          },
          "detail": {
            "type": "string"
          },
          "instance": {
            "type": "string"
          },
          "codigo": {
            "type": "string",
            "description": "Código estável do erro (ex.: prazo.ja-confirmado)."
          },
          "problemas": {
            "description": "Problemas por campo, nos erros de validação.",
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "campo": {
                  "type": "string"
                },
                "mensagem": {
                  "type": "string"
                }
              },
              "required": [
                "campo",
                "mensagem"
              ],
              "additionalProperties": false
            }
          },
          "requestId": {
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "codigo"
        ],
        "additionalProperties": false
      },
...
      "SegundoFatorAtivado": {
        "type": "object",
        "properties": {
          "sessao": {
            "$ref": "#/components/schemas/SessaoAtual"
          },
          "codigosDeRecuperacao": {
            "minItems": 10,
            "maxItems": 10,
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Exibidos só agora: cada um vale uma vez se o celular for perdido."
          }
        },
        "required": [
          "sessao",
          "codigosDeRecuperacao"
        ],
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "sessao": {
        "type": "apiKey",
        "in": "cookie",
        "name": "pz_sessao"
      },
      "oauth2": {
        "type": "http",
        "scheme": "bearer",
        "description": "Token OAuth 2.1 (ADR-015)."
      }
    }
  },
  "security": [
    {
      "sessao": []
    },
    {
      "oauth2": []
    }
  ]
}